src/pyproject.toml
[tool.poetry]
name = "lunchtrack"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
django-model-utils = "^4.3.1"
django-cors-headers = "^4.1.0"
pandas = "^2.0.2"
ipywidgets = "^8.0.6"
workalendar = "^17.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
